Move main to bin module
- Move all methods from swcc module to a common module
- Add swcci binary
