+++ src/lib.rs
@@ -201,10 +201,6 @@ pub struct Opt {
     /// An empty path is allowed but not recommended; it will cause the preprocessor
     /// to look for includes relative to the current directory of the process.
     pub filename: PathBuf,
-
-    #[cfg(feature = "repl")]
-    /// Indicates whether the repl should be started.
-    pub start_repl: bool,
 }
 
 /// Preprocess the source and return the tokens.
